pyproject.toml
[project]
name = "CMS-DB"
version = "0.1.0"
description = "DB for CMS Blog editor"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django==6.0.2",
  "celery==5.6.2",
  "django-debug-toolbar==6.2.0",
  "gunicorn==25.0.1",
  "psycopg==3.3.2",
  "ruff==0.15.0",
  "setuptools==80.10.2",
  "whitenoise==6.11.0",
"djangorestframework==3.15.2",
"django-cors-headers==4.4.0",
"Pillow==10.4.0",
"python-dotenv==1.0.1",
"djangorestframework-stubs==3.15.2",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = ["I", "SIM"]
